[Java] Use long offsets in files for mapping buffer regions in ready flyweights.

+++ aeron-client/src/main/java/uk/co/real_logic/aeron/BufferManager.java
@@ -28,5 +28,5 @@ interface BufferManager
      * @param length   of the region in the file to be mapped.
      * @return a {@link ManagedBuffer} representing the mapped buffer.
      */
-    ManagedBuffer mapBuffer(String fileName, int offset, int length);
+    ManagedBuffer mapBuffer(String fileName, long offset, int length);
 }
